+++ scripts/modules/git/git.sh
@@ -8,7 +8,7 @@ smu::import base
 
 # git functions
 
-clone_git_repo_in() {
+git::clone_git_repo_in() {
 
     TARGET="$1"
     URL="$2"
@@ -19,8 +19,16 @@ clone_git_repo_in() {
 
 }
 
-is_git_repository() {
+git::is_git_repository() {
 
     git rev-parse &> /dev/null
 
 }
+
+# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+# Backwards-compatible aliases (pre-1.3.0 unnamespaced names).
+# New code should call the namespaced functions above.
+
+clone_git_repo_in() { git::clone_git_repo_in "$@"; }
+is_git_repository() { git::is_git_repository "$@"; }