+++ scripts/modules/rust/cargo.sh
@@ -8,7 +8,7 @@ smu::import base
 
 # cargo functions
 
-is_cargo_installed() {
+cargo::is_cargo_installed() {
 
     if ! cmd_exists "cargo"; then
         return 1
@@ -16,7 +16,7 @@ is_cargo_installed() {
 
 }
 
-cargo_install() {
+cargo::cargo_install() {
 
     local package="$1"
 
@@ -33,3 +33,11 @@ cargo_install() {
     fi
 
 }
+
+# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+# Backwards-compatible aliases (pre-1.3.0 unnamespaced names).
+# New code should call the namespaced functions above.
+
+is_cargo_installed() { cargo::is_cargo_installed "$@"; }
+cargo_install() { cargo::cargo_install "$@"; }